+++ nacc_attribute_deriver/attributes/derived/modules/lbd/form_lbd.py
@@ -27,11 +27,13 @@ class LBDFormAttributeCollection(AttributeCollection):
         assumes LBD form."""
         return 1
 
-    ###################################################
-    # Age attributes - aka those with a 777 prev code #
-    ###################################################
+    ################################################
+    # Prev attributes - aka those with a prev code #
+    ################################################
 
-    def __handle_age_attribute(self, field: str, prev_code: int = 777) -> Optional[int]:
+    def __handle_prev_attribute(
+        self, field: str, prev_code: int = 777
+    ) -> Optional[int]:
         """Keep track of non-777s."""
         value = self.__lbd.get_value(field, int)
         if value == prev_code:
@@ -45,31 +47,35 @@ class LBDFormAttributeCollection(AttributeCollection):
 
     def _create_lbpsyage(self) -> Optional[int]:
         """Captures LBPSYAGE."""
-        return self.__handle_age_attribute("lbpsyage")
+        return self.__handle_prev_attribute("lbpsyage")
 
     def _create_lbsagerm(self) -> Optional[int]:
         """Captures LBSAGERM."""
-        return self.__handle_age_attribute("lbsagerm")
+        return self.__handle_prev_attribute("lbsagerm")
 
     def _create_lbsagesm(self) -> Optional[int]:
         """Captures LBSAGESM."""
-        return self.__handle_age_attribute("lbsagesm")
+        return self.__handle_prev_attribute("lbsagesm")
 
     def _create_lbsagegt(self) -> Optional[int]:
         """Captures LBSAGEGT."""
-        return self.__handle_age_attribute("lbsagegt")
+        return self.__handle_prev_attribute("lbsagegt")
 
     def _create_lbsagefl(self) -> Optional[int]:
         """Captures LBSAGEFL."""
-        return self.__handle_age_attribute("lbsagefl")
+        return self.__handle_prev_attribute("lbsagefl")
 
     def _create_lbsagetr(self) -> Optional[int]:
         """Captures LBSAGETR."""
-        return self.__handle_age_attribute("lbsagetr")
+        return self.__handle_prev_attribute("lbsagetr")
 
     def _create_lbsagebr(self) -> Optional[int]:
         """Captures LBSAGEBR."""
-        return self.__handle_age_attribute("lbsagebr")
+        return self.__handle_prev_attribute("lbsagebr")
+
+    def _create_lbspsym(self) -> Optional[int]:
+        """Captures LBSPSYM."""
+        return self.__handle_prev_attribute("lbspsym", prev_code=0)
 
     #######
     # B4l #
@@ -77,27 +83,27 @@ class LBDFormAttributeCollection(AttributeCollection):
 
     def _create_lbdage(self) -> Optional[int]:
         """Captures LBDAGE."""
-        return self.__handle_age_attribute("lbdage")
+        return self.__handle_prev_attribute("lbdage")
 
     def _create_lbdage2(self) -> Optional[int]:
         """Captures LBDAGE2."""
-        return self.__handle_age_attribute("lbdage2")
+        return self.__handle_prev_attribute("lbdage2")
 
     def _create_lbdelage(self) -> Optional[int]:
         """Captures LBDELAGE."""
-        return self.__handle_age_attribute("lbdelage")
+        return self.__handle_prev_attribute("lbdelage")
 
     def _create_lbhalage(self) -> Optional[int]:
         """Captures LBHALAGE."""
-        return self.__handle_age_attribute("lbhalage")
+        return self.__handle_prev_attribute("lbhalage")
 
     def _create_lbanxage(self) -> Optional[int]:
         """Captures LBANXAGE."""
-        return self.__handle_age_attribute("lbanxage")
+        return self.__handle_prev_attribute("lbanxage")
 
     def _create_lbapaage(self) -> Optional[int]:
         """Captures LBAPAAGE."""
-        return self.__handle_age_attribute("lbapaage")
+        return self.__handle_prev_attribute("lbapaage")
 
     #######
     # B9l #
@@ -105,8 +111,12 @@ class LBDFormAttributeCollection(AttributeCollection):
 
     def _create_sccoagen(self) -> Optional[int]:
         """Captures SCCOAGEN."""
-        return self.__handle_age_attribute("sccoagen")
+        return self.__handle_prev_attribute("sccoagen")
 
     def _create_sccoaged(self) -> Optional[int]:
         """Captures SCCOAGED."""
-        return self.__handle_age_attribute("sccoaged")
+        return self.__handle_prev_attribute("sccoaged")
+
+    def _create_sccofrst(self) -> Optional[int]:
+        """Captures SCCOFRST."""
+        return self.__handle_prev_attribute("sccofrst", prev_code=0)