+++ nacc_attribute_deriver/attributes/missingness/modules/uds/missingness_uds.py
@@ -11,6 +11,7 @@ from nacc_attribute_deriver.attributes.collection.uds_collection import (
 from nacc_attribute_deriver.attributes.namespace.keyed_namespace import (
     T,
 )
+from nacc_attribute_deriver.utils.date import standardize_date
 
 
 class GenericUDSMissingness(UDSMissingness):
@@ -19,4 +20,11 @@ class GenericUDSMissingness(UDSMissingness):
 
     def _missingness_uds(self, field: str, attr_type: Type[T]) -> T:
         """Defines general missingness for UDS; -4 if missing."""
+
+        # standardize FRMDATEX variables
+        if field.startswith("frmdate") and attr_type == str:  # noqa: E721
+            formatted_date = standardize_date(self.uds.get_value(field, str))
+            if formatted_date:
+                return formatted_date  # type: ignore
+
         return self.generic_missingness(field, attr_type)