+++ tests/utils/test_utils.py
@@ -2,7 +2,11 @@
 
 import pytest
 
-from nacc_attribute_deriver.utils.date import datetime_from_form_date
+from nacc_attribute_deriver.utils.date import (
+    datetime_from_form_date,
+    find_closest_date,
+    standardize_date,
+)
 from nacc_attribute_deriver.utils.errors import AttributeDeriverError
 
 
@@ -28,3 +32,29 @@ class TestDateUtils:
             == "Failed to parse date 01/01/2025-extrastuff: "
             + "unconverted data remains: -extrastuff"
         )
+
+    def test_standardize_date(self):
+        """Test standardizing date."""
+        assert standardize_date("2025-12-01") == "2025-12-01"
+        assert standardize_date("2025/12/01") == "2025-12-01"
+        assert standardize_date("12-01-2025") == "2025-12-01"
+        assert standardize_date("12/01/2025") == "2025-12-01"
+
+        assert standardize_date(None) is None
+
+    def test_find_closest_date(self):
+        """Test finding the closest date, also mix up the date formats."""
+        # a year after the latest
+        assert find_closest_date(
+            ["2025-01-01", "2026/01/01", "01-01-2027"], "2028-01-01"
+        ) == ("2027-01-01", 2)
+
+        # a day before the middle
+        assert find_closest_date(
+            ["01/01/2025", "2026-01-01", "2027/01/01"], "2025/12/31"
+        ) == ("2026-01-01", 1)
+
+        # exactly the first
+        assert find_closest_date(
+            ["2025/01/01", "01/01/2026", "2027-01-01"], "01-01-2025"
+        ) == ("2025-01-01", 0)